Refactor: 로그인 시 사용자 등록 함수 패키지 이동
