+++ ilkda/src/main/java/com/ilkda/server/auth/service/AuthService.java
@@ -3,10 +3,8 @@ package com.ilkda.server.auth.service;
 import com.ilkda.server.auth.dto.KakaoUserInfo;
 import com.ilkda.server.auth.dto.TokenDTO;
 import com.ilkda.server.exception.UnauthorizedException;
-import com.ilkda.server.member.model.Role;
+import com.ilkda.server.member.service.MemberService;
 import com.ilkda.server.security.provider.JwtProvider;
-import com.ilkda.server.member.model.Member;
-import com.ilkda.server.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
@@ -16,23 +14,21 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
 @RequiredArgsConstructor
 @Service
-@Transactional(readOnly = true)
 public class AuthService {
 
-    private final MemberRepository memberRepository;
     private final JwtProvider jwtProvider;
+    private final MemberService memberService;
 
     public TokenDTO registerUser(String kakaoToken) {
         // 1. Access token으로 카카오에서 사용자 정보 가져오기
         KakaoUserInfo kakaoUserInfo = getKakaoUserInfo(kakaoToken);
 
         // 2. 사용자 정보 저장하기
-        Long memberId = saveUser(kakaoUserInfo);
+        Long memberId = memberService.saveUser(kakaoUserInfo);
 
         // 3. Access token, Refresh token 발급하기
         TokenDTO tokenDTO = new TokenDTO(
@@ -42,14 +38,14 @@ public class AuthService {
         return tokenDTO;
     }
 
-    public KakaoUserInfo getKakaoUserInfo(String accessToken)  {
+    private KakaoUserInfo getKakaoUserInfo(String accessToken)  {
         HttpHeaders headers = new HttpHeaders();
         headers.add("Authorization", String.format("Bearer %s", accessToken));
         HttpEntity request = new HttpEntity(headers);
         RestTemplate restTemplate = new RestTemplate();
         ResponseEntity<String> response;
         try {
-            response = restTemplate.exchange("https://kapi.kakao.com/v2/user/me HTTP/1.1",
+            response = restTemplate.exchange("https://kapi.kakao.com/v2/user/me",
                     HttpMethod.GET,
                     request,
                     String.class);
@@ -58,24 +54,12 @@ public class AuthService {
         }
 
         JSONParser jsonParser = new JSONParser();
-        JSONObject jsonObject = null;
+        JSONObject jsonObject;
         try {
             jsonObject = (JSONObject) jsonParser.parse(response.getBody());
         } catch (ParseException p) {
             throw new UnauthorizedException("토큰 파싱 실패");
         }
         return new KakaoUserInfo(jsonObject);
-    }
-
-    @Transactional
-    public Long saveUser(KakaoUserInfo kakaoUserInfo) {
-        Member member = Member.builder()
-                .kakaoId(kakaoUserInfo.getId())
-                .nickname(kakaoUserInfo.getKakaoAccount().getProfile().getNickname())
-                .profileImage(kakaoUserInfo.getKakaoAccount().getProfile().getProfile_image_url())
-                .role(Role.ROLE_USER)
-                .build();
-        memberRepository.save(member);
-        return member.getId();
     }
 }
